Refactor repository delivery and review into focused internal modules
