+++ packages/agent-repository/src/Agent/CLI/RepositoryDelivery.hs
@@ -19,18 +19,17 @@ import Agent.CLI.RepositoryDelivery.ConfirmationStore
     , newConfirmationStore
     , takeConfirmation
     )
-import Control.Concurrent (threadDelay)
-import Control.Concurrent.Async (cancel, wait, withAsync)
-import Control.Applicative ((<|>))
+import Agent.CLI.RepositoryDelivery.Process
+    ( ProcessResult(..)
+    , ProcessFailure(..)
+    , runCommand
+    , runCommandWithEnvironment
+    , trySynchronous
+    )
+import Agent.CLI.RepositoryDelivery.Validation
 import Control.Exception.Safe
-    ( SomeException
-    , bracket
-    , finally
-    , isAsyncException
-    , mask
+    ( bracket
     , onException
-    , throwIO
-    , tryAny
     )
 import Control.Monad (unless, when)
 import Control.Monad.IO.Class (liftIO)
@@ -44,13 +43,7 @@ import qualified Data.Aeson as Aeson
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Char8 as BS8
 import qualified Data.ByteString.Lazy as LBS
-import Data.Char (isAlphaNum, isHexDigit, isSpace)
-import Data.IORef
-    ( newIORef
-    , readIORef
-    , writeIORef
-    )
-import Data.Maybe (fromMaybe)
+import Data.Char (isHexDigit, isSpace)
 import Data.Text (Text)
 import qualified Data.Text as Text
 import qualified Data.Text.Encoding as TextEncoding
@@ -66,10 +59,8 @@ import System.Directory
     , removePathForcibly
     )
 import System.Exit (ExitCode(..))
-import System.Environment (getEnvironment)
 import System.IO
-    ( Handle
-    , IOMode(ReadMode)
+    ( IOMode(ReadMode)
     , hClose
     , openBinaryFile
     )
@@ -85,24 +76,7 @@ import System.Posix.Files
     , setFileMode
     , unionFileModes
     )
-import System.Posix.Signals
-    ( Signal
-    , sigKILL
-    , sigTERM
-    , signalProcessGroup
-    )
-import System.Posix.Types (ProcessID)
 import System.Posix.Temp (mkdtemp)
-import System.Process
-    ( CreateProcess(..)
-    , ProcessHandle
-    , StdStream(CreatePipe)
-    , createProcess
-    , getPid
-    , proc
-    , terminateProcess
-    , waitForProcess
-    )
 import System.Timeout (timeout)
 
 import Agent.CLI.RepositoryReview
@@ -112,7 +86,6 @@ import Agent.CLI.RepositoryReview
 import Agent.CLI.ProcessSecurity
     ( canonicalPathOutside
     , resolveExecutableOutside
-    , sanitizeSearchPathOutside
     )
 
 data DeliveryStatus = DeliveryStatus
@@ -198,19 +171,6 @@ data StoredConfirmation = StoredConfirmation
     { storedRoot :: !FilePath
     , storedConfirmation :: !Confirmation
     }
-
-data ProcessResult = ProcessResult
-    { processExitCode :: !ExitCode
-    , processStdout :: !BS.ByteString
-    , processStderr :: !BS.ByteString
-    , processOutputTruncated :: !Bool
-    }
-
-data ProcessFailure
-    = ProcessLaunchFailed
-    | ProcessTimedOut
-    | ProcessOutputExceeded
-    deriving (Eq, Show)
 
 repositoryDeliveryStatus
     :: FilePath
@@ -1117,114 +1077,6 @@ pullRequestNumber repository url =
             | number > 0 -> Just number
         _ -> Nothing
 
-validRemoteUrl :: BS.ByteString -> Bool
-validRemoteUrl url =
-    not (BS.null url)
-        && BS.length url <= 4096
-        && BS.all (\byte -> byte >= 0x20 && byte /= 0x7f) url
-        && BS8.head url /= '-'
-        && (validHttps url
-            || validSsh url
-            || validScpLike url)
-  where
-    validHttps value =
-        "https://" `BS8.isPrefixOf` value
-            && validUrlAuthority "https://" value
-            && not (authorityContains '@' "https://" value)
-            && not (containsQueryOrFragment value)
-            && not (authorityContains '%' "https://" value)
-    validSsh value =
-        "ssh://" `BS8.isPrefixOf` value
-            && validSshAuthority value
-            && not (containsQueryOrFragment value)
-    authorityContains character prefix value =
-        BS8.elem character
-            (BS8.takeWhile (/= '/') (BS.drop (BS.length prefix) value))
-    validSshAuthority value =
-        let authority =
-                BS8.takeWhile (/= '/') (BS.drop (BS.length "ssh://") value)
-            (userinfo, separatorAndHost) = BS8.break (== '@') authority
-            host = BS.drop 1 separatorAndHost
-        in not (BS.null authority)
-            && not (BS8.elem '%' authority)
-            && if BS.null separatorAndHost
-                then validHostPort authority
-                else validUsername userinfo
-                    && validHostPort host
-                    && not (BS8.elem '@' host)
-    validUrlAuthority prefix value =
-        validHostPort
-            (BS8.takeWhile (/= '/') (BS.drop (BS.length prefix) value))
-    validHostPort authority =
-        case BS8.break (== ':') authority of
-            (host, port)
-                | BS.null port -> validHost host
-                | otherwise ->
-                    validHost host
-                        && BS.length port > 1
-                        && BS8.all
-                            (\character ->
-                                character >= '0' && character <= '9')
-                            (BS.drop 1 port)
-    validHost host =
-        not (BS.null host)
-            && BS8.all
-                (\character ->
-                    isAsciiAlphaNumeric character
-                        || character `elem` (".-" :: String))
-                host
-            && BS8.head host /= '.'
-            && BS8.last host /= '.'
-    validUsername username =
-        not (BS.null username)
-            && BS8.all
-                (\character ->
-                    isAsciiAlphaNumeric character
-                        || character `elem` ("._-" :: String))
-                username
-    isAsciiAlphaNumeric character =
-        (character >= 'a' && character <= 'z')
-            || (character >= 'A' && character <= 'Z')
-            || (character >= '0' && character <= '9')
-    containsQueryOrFragment value =
-        BS8.elem '?' value || BS8.elem '#' value
-    validScpLike value =
-        case BS8.break (== ':') value of
-            (authority, path) ->
-                let (username, separatorAndHost) = BS8.break (== '@') authority
-                    host = BS.drop 1 separatorAndHost
-                in validUsername username
-                    && not (BS.null separatorAndHost)
-                    && validHost host
-                    && not (BS8.elem '@' host)
-                    && not (BS8.elem '%' authority)
-                    && BS.length path > 1
-                    && not (BS8.elem '@' path)
-                    && not (containsQueryOrFragment value)
-
-githubRepositoryFromUrl :: BS.ByteString -> Maybe Text
-githubRepositoryFromUrl bytes
-    | not (BS.all (< 0x80) bytes) = Nothing
-    | otherwise =
-        let url = Text.pack (BS8.unpack bytes)
-        in parseHttps url
-            <|> parseSsh url
-            <|> parseScp url
-  where
-    parseHttps url =
-        Text.stripPrefix "https://github.com/" url >>= repositoryPath
-    parseSsh url =
-        (Text.stripPrefix "ssh://git@github.com/" url
-            <|> Text.stripPrefix "ssh://github.com/" url)
-            >>= repositoryPath
-    parseScp url =
-        Text.stripPrefix "git@github.com:" url >>= repositoryPath
-    repositoryPath path =
-        let withoutGit = fromMaybe path (Text.stripSuffix ".git" path)
-        in if validateRepositoryName withoutGit
-            then Just withoutGit
-            else Nothing
-
 remoteMatchesExpected :: DeliveryStatus -> Maybe Text -> Bool
 remoteMatchesExpected status =
     (== expectedRemoteHead status)
@@ -1282,75 +1134,6 @@ validatePullRequestInput base title body
     | Text.any (== '\NUL') body =
         Left (DeliveryInvalidRequest "pull-request body contains a NUL byte")
     | otherwise = Right ()
-
-validateBranchName :: Text -> Bool
-validateBranchName branch =
-    validateFullBranchRef ("refs/heads/" <> branch)
-
-validateFullBranchRef :: Text -> Bool
-validateFullBranchRef ref =
-    not (Text.null ref)
-        && Text.length ref <= 1024
-        && "refs/heads/" `Text.isPrefixOf` ref
-        && not ("/" `Text.isSuffixOf` ref)
-        && not ("." `Text.isSuffixOf` ref)
-        && not ("." `Text.isPrefixOf` ref)
-        && not ("-" `Text.isPrefixOf` Text.drop (Text.length "refs/heads/") ref)
-        && not (".." `Text.isInfixOf` ref)
-        && not ("@{" `Text.isInfixOf` ref)
-        && not ("//" `Text.isInfixOf` ref)
-        && Text.all safeRefCharacter ref
-        && all validComponent (Text.splitOn "/" ref)
-  where
-    safeRefCharacter character =
-        not (isSpace character)
-            && character >= '\x20'
-            && character /= '\x7f'
-            && character `notElem` ("~^:?*[\\" :: String)
-    validComponent component =
-        not (Text.null component)
-            && not ("." `Text.isPrefixOf` component)
-            && component /= "."
-            && component /= ".."
-            && not (".lock" `Text.isSuffixOf` component)
-
-validateRemoteName :: Text -> Bool
-validateRemoteName remote =
-    not (Text.null remote)
-        && Text.length remote <= 255
-        && Text.head remote /= '-'
-        && Text.all
-            (\character ->
-                isAlphaNum character
-                    || character `elem` ("._/-" :: String))
-            remote
-        && not (".." `Text.isInfixOf` remote)
-        && not ("//" `Text.isInfixOf` remote)
-
-validateRepositoryName :: Text -> Bool
-validateRepositoryName name =
-    case Text.splitOn "/" name of
-        [owner, repository] ->
-            validPart owner && validPart repository
-        _ -> False
-  where
-    validPart value =
-        not (Text.null value)
-            && value /= "."
-            && value /= ".."
-            && Text.length value <= 100
-            && Text.all
-                (\character ->
-                    isAlphaNum character
-                        || character `elem` ("-._" :: String))
-                value
-
-validObjectId :: Text -> Bool
-validObjectId oid =
-    Text.length oid `elem` [40, 64] && Text.all isHexDigit oid
-
-zeroObjectId :: Text -> Text
-zeroObjectId oid = Text.replicate (Text.length oid) "0"
 
 parseAheadBehind :: BS.ByteString -> Maybe (Int, Int)
 parseAheadBehind bytes =
@@ -1848,342 +1631,12 @@ runGh root arguments input timeoutMicros = do
                         (DeliveryUnavailable
                             "GitHub CLI command failed"))
 
-runCommand
-    :: FilePath
-    -> FilePath
-    -> [String]
-    -> BS.ByteString
-    -> Int
-    -> IO (Either ProcessFailure ProcessResult)
-runCommand root =
-    runCommandWithEnvironment [] root root
-
-runCommandWithEnvironment
-    :: [(String, String)]
-    -> FilePath
-    -> FilePath
-    -> FilePath
-    -> [String]
-    -> BS.ByteString
-    -> Int
-    -> IO (Either ProcessFailure ProcessResult)
-runCommandWithEnvironment
-    overrides
-    trustRoot
-    workingDirectory
-    executable
-    arguments
-    input
-    timeoutMicros = do
-    launched <- trySynchronous
-        (bracket start stop \processData ->
-            timeout timeoutMicros (run processData))
-    pure case launched of
-        Left _ -> Left ProcessLaunchFailed
-        Right Nothing -> Left ProcessTimedOut
-        Right (Just result) -> result
-  where
-    start = mask \_ -> do
-        resolvedExecutable <-
-            resolveExecutableOutside trustRoot executable
-                >>= either (fail . Text.unpack) pure
-        environment <- applyEnvironmentOverrides overrides
-            <$> nonInteractiveEnvironment trustRoot executable
-        (maybeInput, maybeOutput, maybeError, process) <-
-            createProcess
-                (proc resolvedExecutable arguments)
-                    { cwd = Just workingDirectory
-                    , std_in = CreatePipe
-                    , std_out = CreatePipe
-                    , std_err = CreatePipe
-                    , close_fds = True
-                    , create_group = True
-                    , env = Just environment
-                    }
-        case (maybeInput, maybeOutput, maybeError) of
-            (Just inputHandle, Just outputHandle, Just errorHandle) -> do
-                let closePipes = do
-                        closeQuietly inputHandle
-                        closeQuietly outputHandle
-                        closeQuietly errorHandle
-                    cleanupWithoutGroup = do
-                        closePipes
-                        _ <- tryAny (terminateProcess process)
-                        _ <- tryAny (waitForProcess process)
-                        pure ()
-                processGroup <- getPid process
-                    `onException` cleanupWithoutGroup
-                completed <- newIORef False
-                    `onException` do
-                        closePipes
-                        terminateProcessGroup sigKILL processGroup process
-                        _ <- tryAny (waitForProcess process)
-                        pure ()
-                pure
-                    ( inputHandle
-                    , outputHandle
-                    , errorHandle
-                    , process
-                    , processGroup
-                    , completed
-                    )
-            _ -> do
-                terminateProcess process
-                _ <- waitForProcess process
-                fail "could not create command pipes"
-    stop
-        ( inputHandle
-        , outputHandle
-        , errorHandle
-        , process
-        , processGroup
-        , completed
-        ) = do
-        closeQuietly inputHandle
-        closeQuietly outputHandle
-        closeQuietly errorHandle
-        finished <- readIORef completed
-        unless finished do
-            terminateProcessGroup sigTERM processGroup process
-            threadDelay 100_000
-            -- A descendant can retain the group and pipes after its leader
-            -- exits, so always escalate the captured process group.
-            terminateProcessGroup sigKILL processGroup process
-        _ <- tryAny (waitForProcess process)
-        pure ()
-    run
-        ( inputHandle
-        , outputHandle
-        , errorHandle
-        , process
-        , processGroup
-        , completed
-        ) =
-        withAsync
-            (BS.hPut inputHandle input `finally` closeQuietly inputHandle)
-            \inputWriter ->
-                withAsync (readBounded outputHandle) \outputReader ->
-                    withAsync (readBounded errorHandle) \errorReader -> do
-                        exitCode <- waitForProcess process
-                        inputFinished <- timeout processPipeTeardownMicros
-                            (wait inputWriter)
-                        case inputFinished of
-                            Just () -> pure ()
-                            Nothing -> do
-                                closeQuietly inputHandle
-                                cancel inputWriter
-                        let drainReaders =
-                                (,) <$> wait outputReader <*> wait errorReader
-                        -- Give ordinary buffered output a short chance to
-                        -- drain. The captured group is then terminated even
-                        -- when EOF already arrived: a descendant may close
-                        -- its pipes while continuing to run.
-                        naturallyDrained <- timeout
-                            processPipeTeardownMicros
-                            drainReaders
-                        terminateProcessGroup sigTERM processGroup process
-                        drainedAfterTerm <- case naturallyDrained of
-                            Just values -> pure (Just values)
-                            Nothing ->
-                                timeout processGroupTermGraceMicros drainReaders
-                        -- Always escalate the captured group so a descendant
-                        -- cannot outlive a successful leader.
-                        terminateProcessGroup sigKILL processGroup process
-                        drained <- case drainedAfterTerm of
-                            Just values -> pure (Just values)
-                            Nothing ->
-                                timeout processPipeTeardownMicros drainReaders
-                        case drained of
-                            Nothing -> do
-                                closeQuietly outputHandle
-                                closeQuietly errorHandle
-                                cancel outputReader
-                                cancel errorReader
-                                pure (Left ProcessTimedOut)
-                            Just
-                                ( (output, outputTruncated)
-                                , (errors, errorsTruncated)
-                                ) -> do
-                                    writeIORef completed True
-                                    let truncated =
-                                            outputTruncated || errorsTruncated
-                                    pure
-                                        (if truncated
-                                            then Left ProcessOutputExceeded
-                                            else
-                                                Right
-                                                    ProcessResult
-                                                        { processExitCode =
-                                                            exitCode
-                                                        , processStdout = output
-                                                        , processStderr = errors
-                                                        , processOutputTruncated =
-                                                            False
-                                                        })
-
-applyEnvironmentOverrides
-    :: [(String, String)]
-    -> [(String, String)]
-    -> [(String, String)]
-applyEnvironmentOverrides overrides inherited =
-    overrides
-        <> filter
-            (\(name, _) -> name `notElem` map fst overrides)
-            inherited
-
-data BoundedReadState = BoundedReadState
-    { boundedChunks :: ![BS.ByteString]
-    , boundedRetainedBytes :: !Int
-    , boundedTruncated :: !Bool
-    }
-
-emptyBoundedReadState :: BoundedReadState
-emptyBoundedReadState = BoundedReadState
-    { boundedChunks = []
-    , boundedRetainedBytes = 0
-    , boundedTruncated = False
-    }
-
-retainBoundedChunk :: BoundedReadState -> BS.ByteString -> BoundedReadState
-retainBoundedChunk state chunk =
-    BoundedReadState
-        { boundedChunks =
-            if BS.null kept
-                then state.boundedChunks
-                else kept : state.boundedChunks
-        , boundedRetainedBytes =
-            state.boundedRetainedBytes + BS.length kept
-        , boundedTruncated =
-            state.boundedTruncated || BS.length kept < BS.length chunk
-        }
-  where
-    room = max 0 (maxProcessOutputBytes - state.boundedRetainedBytes)
-    kept = BS.take room chunk
-
-readBounded :: Handle -> IO (BS.ByteString, Bool)
-readBounded handle = do
-    finalState <-
-        drain emptyBoundedReadState `finally` closeQuietly handle
-    pure
-        ( BS.concat (reverse finalState.boundedChunks)
-        , finalState.boundedTruncated
-        )
-  where
-    drain state = do
-        chunk <- BS.hGetSome handle (64 * 1024)
-        if BS.null chunk
-            then pure state
-            else do
-                let nextState = retainBoundedChunk state chunk
-                nextState `seq` drain nextState
-
-terminateProcessGroup
-    :: Signal
-    -> Maybe ProcessID
-    -> ProcessHandle
-    -> IO ()
-terminateProcessGroup signal processGroup process = do
-    pid <- maybe (getPid process) (pure . Just) processGroup
-    case pid of
-        Nothing -> do
-            _ <- tryAny (terminateProcess process)
-            pure ()
-        Just processId -> do
-            _ <- tryAny (signalProcessGroup signal processId)
-            pure ()
-
 stripLineEnding :: BS.ByteString -> BS.ByteString
 stripLineEnding = BS8.dropWhileEnd (`elem` ['\r', '\n'])
 
 decodeTrimmed :: BS.ByteString -> Text
 decodeTrimmed =
     Text.strip . TextEncoding.decodeUtf8With lenientDecode
-
-closeQuietly :: Handle -> IO ()
-closeQuietly handle = do
-    _ <- tryAny (hClose handle)
-    pure ()
-
-nonInteractiveEnvironment :: FilePath -> FilePath -> IO [(String, String)]
-nonInteractiveEnvironment root executable = do
-    inherited <- getEnvironment
-    let blocked =
-            [ "GIT_TERMINAL_PROMPT"
-            , "GCM_INTERACTIVE"
-            , "GH_PROMPT_DISABLED"
-            , "GH_REPO"
-            , "GH_HOST"
-            , "SSH_ASKPASS_REQUIRE"
-            , "GIT_DIR"
-            , "GIT_WORK_TREE"
-            , "GIT_INDEX_FILE"
-            , "GIT_OBJECT_DIRECTORY"
-            , "GIT_ALTERNATE_OBJECT_DIRECTORIES"
-            , "GIT_COMMON_DIR"
-            , "GIT_CONFIG_COUNT"
-            , "GIT_CONFIG_KEY_0"
-            , "GIT_CONFIG_VALUE_0"
-            , "GIT_CONFIG_PARAMETERS"
-            , "GIT_CONFIG_GLOBAL"
-            , "GIT_CONFIG_SYSTEM"
-            , "GIT_CONFIG_NOSYSTEM"
-            , "GIT_ATTR_NOSYSTEM"
-            , "GIT_CEILING_DIRECTORIES"
-            , "GIT_SSH_COMMAND"
-            , "GIT_ASKPASS"
-            , "GIT_PROXY_COMMAND"
-            , "GIT_EXEC_PATH"
-            ]
-        sanitized = filter
-            (\(name, _) ->
-                name `notElem` blocked
-                    && not ("GIT_CONFIG_KEY_" `prefixOf` name)
-                    && not ("GIT_CONFIG_VALUE_" `prefixOf` name))
-            inherited
-    safePath <-
-        case lookup "PATH" sanitized of
-            Nothing -> pure Nothing
-            Just value -> sanitizeSearchPathOutside root value
-    let
-        sanitizedWithSafePath =
-            case safePath of
-                Nothing -> filter ((/= "PATH") . fst) sanitized
-                Just value ->
-                    ("PATH", value) : filter ((/= "PATH") . fst) sanitized
-        retained
-            | executable == "git" =
-                filter (\(name, _) -> name `elem` gitEnvironmentAllowlist)
-                    sanitizedWithSafePath
-            | otherwise = sanitizedWithSafePath
-    pure
-        ( [ ("GIT_TERMINAL_PROMPT", "0")
-          , ("GCM_INTERACTIVE", "never")
-          , ("GH_PROMPT_DISABLED", "true")
-          , ("SSH_ASKPASS_REQUIRE", "never")
-          ]
-            <> retained
-        )
-  where
-    prefixOf prefix value = take (length prefix) value == prefix
-    gitEnvironmentAllowlist =
-        [ "PATH"
-        , "HOME"
-        , "TMPDIR"
-        , "TMP"
-        , "TEMP"
-        , "LANG"
-        , "LC_ALL"
-        , "LC_CTYPE"
-        , "USER"
-        , "LOGNAME"
-        , "SSH_AUTH_SOCK"
-        , "XDG_CONFIG_HOME"
-        , "XDG_CONFIG_DIRS"
-        , "SSL_CERT_FILE"
-        , "SSL_CERT_DIR"
-        , "NIX_SSL_CERT_FILE"
-        , "TERM"
-        ]
 
 deliveryErrorText :: DeliveryError -> Text
 deliveryErrorText = \case
@@ -2192,14 +1645,6 @@ deliveryErrorText = \case
     DeliveryUnavailable message -> message
     DeliveryCommandFailed message -> message
     DeliveryConfirmationRejected message -> message
-
-trySynchronous :: IO value -> IO (Either SomeException value)
-trySynchronous action =
-    tryAny action >>= \case
-        Left exception
-            | isAsyncException exception -> throwIO exception
-            | otherwise -> pure (Left exception)
-        Right value -> pure (Right value)
 
 confirmationLifetimeSeconds :: POSIXTime
 confirmationLifetimeSeconds = 10 * 60
@@ -2217,15 +1662,6 @@ localTimeoutMicros = 15 * 1_000_000
 
 networkTimeoutMicros :: Int
 networkTimeoutMicros = 60 * 1_000_000
-
-processPipeTeardownMicros :: Int
-processPipeTeardownMicros = 1_000_000
-
-processGroupTermGraceMicros :: Int
-processGroupTermGraceMicros = 2_000_000
-
-maxProcessOutputBytes :: Int
-maxProcessOutputBytes = 1024 * 1024
 
 maxActiveConfirmations :: Int
 maxActiveConfirmations = 1024
